% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probpool_functions.R
\name{disppool}
\alias{disppool}
\title{Create a probabilistic dispersal pool}
\usage{
disppool(disp_ability, occurrences, conductance, method = c("negexp",
  "fattail"), longlat = TRUE)
}
\arguments{
\item{disp_ability}{A numeric vector of dispersal abilities. [WHAT IS THE UNIT HERE?]}

\item{conductance}{Optional. A Raster* object representing landscape conductance for dispersal. Takes values between 0 and 1, where 0 indicates complete resistance (\eqn{P(dispersal to raster cell) = 0}) and 1 indicates complete conductance (\eqn{P(dispersal to raster cell) = p}). If a single RasterLayer is provided, the conductance matrix will be recycled for each species.}

\item{method}{A number specifying the kernel used for estimating dispersal distance. (see details)}

\item{longlat}{Logical. Should calculations be based on long/lat?}

\item{occurences}{A Raster* object containing occurrence data for individual species. Each species is represented by a binary RasterLayer, where cell values of 0 indicate absence and 1 indicates presence.}
}
\value{
A raster stack containing species-specific probabilities of dispersal from areas of occurrence to a certain raster cell within a given time frame.
}
\description{
Create a probabilistic dispersal pool
}
\seealso{
\code{\link{probpool}}
#' @export
}
