% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_disp_ability}
\alias{example1_disp_ability}
\title{Dispersal abilities example}
\format{rasterBrick}
\usage{
example1_disp_ability
}
\description{
DESCRIPTION HERE
}
\keyword{datasets}
