% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probpool_class.R
\docType{methods}
\name{plot,Probpool,ANY-method}
\alias{plot,Probpool,ANY-method}
\title{Plot a Probpool object.}
\usage{
\S4method{plot}{Probpool,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{Probpool}}

\item{...}{Additional arguments for \code{\link[lattice]{lattice::levelplot}} or \code{\link[lattice]{lattice::xyplot}}}

\item{focal_species}{The name of a particular species in x. If provided, probability pools for that ...}

\item{focal_unit}{An object comapatible with the \code{\link[raster]{raster::extract}} function, i.e 
\itemize{
  \item points represented by a two-column matrix or data.frame
  \item a numeric vector representing cell numbers
  \item a SpatialPoints*; SpatialPolygons*; SpatialLines object
  \item an \code{\link[raster]{raster::Extent}} object
}}
}
\value{
None
}
\description{
Method to plot and visualize a \code{Probpool} object. Depending on the arguments provided, different plotting behaviour (see Details). Additional parameters are passed to the  the '...' argument
}
\examples{
my_prob_pool = probpool(env_pool = env, disp_pool = disp)
plot(my_prob_pool)
plot(my_prob_pool, focal_species = "Olea europaea")
plot(my_prob_pool, focal_unit = 132)) # Cell number
plot(my_prob_pool, focal_unit = c(24,26))) # Cell index
plot(my_prob_pool, focal_unit = extent(c(7,8,49,53))) # Object of class raster::Extent
}
