% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ranunculaceae}
\alias{Ranunculaceae}
\title{Dataset Ranunculaceae}
\format{list}
\usage{
Ranunculaceae
}
\description{
Example data from Karger et al (2016), containing different datasets for 51 species of Ranunculaceae in Germany. [THIS IS STILL VERY INCOMPLETE!]
}
\details{
\describe{
  \item{Ran_env}{Environmental suitability based on [???]}
  \item{Ran_disp}{Dispersal probabilities based on [???]}
  \item{Ran_occ}{Presence/absence data taken from [???]}
  \item{Ran_int}{A matrix of pairwise biotic interactions. Interaction values were obtained using the \pkg{cooccur} package}
  \item{Ran_disp_dist}{Approximate dispersal ability of species based on [???]}
}
}
\references{
Karger, D. N. et al. (2016), Delineating probabilistic species pools in ecology and biogeography. Global Ecology and Biogeography, 25: 489-501. doi:10.1111/geb.12422
}
\keyword{datasets}
