% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probpool_class.R
\name{summary.probpool}
\alias{summary.probpool}
\title{Summarize a Probpool object.}
\usage{
summary.probpool(object, ...)
}
\arguments{
\item{x}{An object of class \code{Probpool}}
}
\value{
A \code{list} of properties of \code{object}
}
\description{
Summary functions for an object of class Probpool. Prints and returns a \code{list} of properties for an object of class \code{Probpool}
}
