% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probpool_class.R
\name{print}
\alias{print}
\alias{summary}
\title{Summarizing a Probpool object}
\usage{
print(x)

summary(x)
}
\arguments{
\item{x}{An object of class \code{Probpool}}
}
\value{
A \code{list} of properties of \code{object}
}
\description{
Prints and returns a \code{list} of properties for an object of class \code{Probpool}
}
