% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probpool_functions.R
\name{probpool}
\alias{probpool}
\title{Create a probabilistic species pool}
\usage{
probpool(env_pool = NULL, disp_pool = NULL, occurrences = NULL,
  interaction_matrix = NULL, interaction_method = 1)
}
\arguments{
\item{env_pool}{A Raster* object containing environmental probabilities}

\item{disp_pool}{A Raster* object containing dispersal probabilies}

\item{interaction_matrix}{An interaction matrix}

\item{interaction_method}{A number specifying the method used for incorporating species interacitons in the calculation of the probabilistic species pool. (see details)}

\item{occurences}{A Raster* object containing occurrence data for individual species. Each species is represented by a binary RasterLayer, where cell values of 0 indicate absence and 1 indicates presence.}
}
\value{
A raster stack containing species-specific probabilities of occurrence
}
\description{
Description goes here
}
\seealso{
\code{\link{Probpool}}
#' @export
}
